+++ opencti-platform/opencti-graphql/config/schema/opencti.graphql
@@ -2122,7 +2122,7 @@ type Connector implements BasicObject & InternalObject {
   manager_connector_logs: [String!]
   manager_contract_hash: String
   manager_health_metrics: ConnectorHealthMetrics
-  manager_connector_uptime: Float
+  manager_connector_uptime: Int
   built_in: Boolean
   configurations: [ConnectorConfiguration!]
 }
@@ -2146,7 +2146,7 @@ type ManagedConnector implements BasicObject & InternalObject {
   manager_contract_hash: String!
   manager_connector_logs: [String!]!
   manager_health_metrics: ConnectorHealthMetrics
-  manager_connector_uptime: Float
+  manager_connector_uptime: Int
 }
 
 type ConnectorHealthMetrics {
